fix win32 caret blink - we don't blink now if win32 says not to
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@12558 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/af/gr/xp/gr_Caret.cpp
@@ -214,7 +214,7 @@ void GR_Caret::setBlink(bool bBlink)
 	g_object_get (G_OBJECT(settings), "gtk-cursor-blink", &can, NULL);
 	m_bCursorBlink = (can != FALSE);
 #elif defined(WIN32)
-	m_bCursorBlink = (GetCaretBlinkTime () > 0);
+	m_bCursorBlink = (((int)GetCaretBlinkTime ()) > 0);
 #else
 	m_bCursorBlink = bBlink;
 #endif
